refactor: reorganized folders and applied snake_case to filenames
