+++ frontend/sustenta_bag_application/lib/main.dart
@@ -1,23 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:sustenta_bag_application/AppShell.dart' show AppShell;
-import 'package:sustenta_bag_application/screens/Edit/EditStep1.dart';
-import 'package:sustenta_bag_application/screens/Edit/EditStep2.dart';
-import 'package:sustenta_bag_application/screens/Review/ShowReviewScreen.dart';
-import 'package:sustenta_bag_application/screens/StoreScreen.dart';
+import 'package:sustenta_bag_application/screens/app_shell.dart' show AppShell;
+import 'package:sustenta_bag_application/screens/Edit/edit_step1.dart';
+import 'package:sustenta_bag_application/screens/Edit/edit_step2.dart';
+import 'package:sustenta_bag_application/screens/Review/show_review_screen.dart';
+import 'package:sustenta_bag_application/screens/store_screen.dart';
 import 'package:sustenta_bag_application/firebase_options.dart';
-import 'package:sustenta_bag_application/screens/UserDataScreen.dart';
+import 'package:sustenta_bag_application/screens/user_data_screen.dart';
 import 'package:sustenta_bag_application/services/firebase_messaging_service.dart';
-import 'screens/IntroScreen.dart';
-import 'screens/LoginScreen.dart';
-import 'screens/Register/RegisterStep1.dart';
-import 'screens/Register/RegisterStep2.dart';
-import 'screens/Register/RegisterStep3.dart';
-import 'screens/bag/DeliveryOptionsScreen.dart';
-import 'screens/bag/ReviewOrderScreen.dart';
-import 'screens/bag/PaymentScreen.dart';
-import 'screens/bag/PendingOrderDetailsScreen.dart';
-import 'screens/FavoritesScreen.dart';
+import 'screens/intro_screen.dart';
+import 'screens/login_screen.dart';
+import 'screens/Register/register_step1.dart';
+import 'screens/Register/register_step2.dart';
+import 'screens/Register/register_step3.dart';
+import 'screens/bag/delivery_options_screen.dart';
+import 'screens/bag/review_order_screen.dart';
+import 'screens/bag/payment_screen.dart';
+import 'screens/bag/pending_order_details_screen.dart';
+import 'screens/favorites_screen.dart';
 import 'models/order.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
