+++ frontend/sustenta_bag_application/lib/screens/home_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../components/bagCard.dart';
+import '../components/bag_card.dart';
 import '../services/firebase_messaging_service.dart';
 import '../services/location_service.dart';
 import '../utils/database_helper.dart';
